move generate_index related code into a separate mod
